# Avoid false-positive warnings about method calls on the Fabric-wrapped module