## What does this PR do?

This PR implements a better way to warn users in case they call a method on the wrapper. 

A warning is currently printed on any call to a method that is not forward(), given that the original module is wrapped by a distributed wrapper like DDP, FSDP or similar. See an example in https://github.com/Lightning-AI/lit-gpt/issues/641
This PR restricts this warning to only the case when an actual module forward call occurs. A call to a method that does not run inputs through the submodules is not considered problematic anymore.

I ran a quick benchmark to measure the impact of adding the model hooks:
```py
import lightning as L
import torch
from lightning.pytorch.demos import Transformer


class NewTransformer(Transformer):
    def new_forward(self, input, target):
        return self(input, target)


def run():
    fabric = L.Fabric(accelerator="cuda", devices=1, strategy="ddp")
    fabric.seed_everything(1)
    fabric.launch()
    model = NewTransformer(ninp=2, nhead=1, nhid=2, nlayers=1, vocab_size=4)
    optimizer = torch.optim.SGD(model.parameters(), lr=0.1)
    model, optimizer = fabric.setup(model, optimizer)
    model.eval()
    input = torch.randint(0, 4, (1, 35), device=fabric.device)
    target = torch.randint(0, 4, (1, 35), device=fabric.device)
    with torch.inference_mode():
        for _ in range(100):
            model.new_forward(input, target)


if __name__ == "__main__":
    run()



```
Times measured with:
```
python -m timeit -r 10 -n 10  -vv  'from speed_test import run; run()'
```

**GPU benchmark**
master:
10 loops, best of 10: 184.9 msec per loop
branch:
10 loops, best of 10: 208.9 msec per loop

**CPU benchmark**
master:
10 loops, best of 10: 151 msec per loop
branch:
10 loops, best of 10: 173.7 msec per loop

The benchmarks show a ~20 msec impact when running with the hooks in this extreme example of a super small model. On regular-sized models, the impact is not measurable. 

<!-- readthedocs-preview pytorch-lightning start -->
----
:books: Documentation preview :books:: https://pytorch-lightning--18819.org.readthedocs.build/en/18819/

<!-- readthedocs-preview pytorch-lightning end -->

cc @borda @carmocca @justusschock @awaelchli