@ardywibowo 2023-11-29
This seems to be breaking `model.generate(...)` for HuggingFace models. Any ideas?

```
RuntimeError: You are calling the method `BloomForCausalLM.generate()` from outside the model. This will bypass the wrapper from the strategy and result in incorrect behavior in `.backward()`. You should pass your inputs through `forward()`.
```